tweak: display smesher id for any node event that contain it
